requirements.txt
flask==2.3.2
flask-mysqldb==1.0.1
pyjwt==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
pymysql==1.1.0
gunicorn==21.2.0